Experimental: Handle incoming links in TransformLinkChange
We found that bidirectional link relationshups such as the Bonfire
testing scenario ("discovered while testing"/"testing
discovered") were not handled completely. The link is added
to only one of the two work items in Azure. The information
required to set up the bidirectional link is already lost
while exporting from JIRA (i.e. in jira-export).

Consider the following scenario in JIRA:
- FOOBAR-4812 has a link to FOOBAR-4607 of type "discovered while testing"
- FOOBAR-4607 has a link to FOOBAR-4812 of type "testing discovered"

After exporting from JIRA, only FOOBAR-4607.json receives link information,
but not FOOBAR-4812.json. (To be more precise, it depends on the export
sequence which of the two .json files is affected.)